Disable LB_LINUX_PATH from within LB_PROG_LINUX
The LB_LINUX_PATH macro was commented out and replaced mainly by the
LDISKFS_AC_LINUX* suite of autoconf macros. The functionality that is
not preserved by the LDISKFS_AC_LINUX* macros but is still needed, was
grouped into the ldiskfs-build-compat.m4 file using the
LDISKFS_AC_COMPAT_LB_* suite of macros. This new file is meant to hold
temporary macros that are needed to hold the build system together
during this refactoring.