Remove dependency on LB_CONFIG_FILES macro
The LB_CONFIG_FILES macro was removed from configure.ac and replace with
an explicit call to AC_CONFIG_FILES. In addition, the build/Rules.in
files was moved out of the build directory and into the root directory
of the package. This keeps the new build infrastructure free from
dependencies in the build directory.