Remove remaining disabled configure.ac features
The remaining features of the configure.ac script that were disabled
have been stripped out by this commit. The remaining portions of the
configure.ac script are what has been deemed necessary to build the
package. As a result, the dependency on the build/autoconf/*.m4 macros
was removed from the autogen.sh script.