Add initial bits of new make infrastructure
The first initial parts of the new make infrastructure were added by
this commit. The core of code contained in the ldiskfs/Makefile.in file
was pulled from the previous build infrastructure to copy and patch the
ext* kernel source files into ldiskfs source files. Rules to actually
compile the ldiskfs source files still need to be added.