Fix the file hierarchy in the devel package
The header files in the devel package were not being placed in the
correct locations. The subdirectories of the ldiskfs directory that
need to be included in the devel package should not be packaged as
subdirectories. Rather, those directories need to packaged along side
the ldiskfs directory. To illustrate, here is an example of the previous
hierarchy:

        ldiskfs/file.h
        ldiskfs/linux/file.h
        ldiskfs/events/trace/file.h

And here is the hierarchy now:

        ldiskfs/file.h
        linux/file.h
        events/trace/file.h