Properly detect file name for symbol versions
To build ldiskfs using the ext sources in Red Hat's
kernel-debuginfo-common package, we point --with-linux at its
installation directory and --with-linux-obj at the kernel-devel
directory.  This was breaking the configure check for the symbol
versions file name, which depends on the file
$LINUX/scripts/Makefile.modpost.  Since that file is provided by
kernel-devel and not kernel-debuginfo-common, we need to look for it
in both the $LINUX and $LINUX_OBJ directories.