Support RHEL6 kernel-debuginfo-common package
Modified the LDISKFS_AC_LINUX autoconf script in an attempt to support
the RHEL6 kernel-debuginfo-common package. That package places the
kernel source files in a subdirectory of /usr/src/debug. The problem is
that the /lib/modules/$(uname -r)/source symlink does not get updated to
point the this directory containing the source files. Thus, this commit
adds a hack to check the debug directory before using the source symlink
to find the kernel source directory. Ideally, the RHEL6
kernel-debuginfo-common package should update the would symlink
correctly.