Add kernel-devel package hint on UTS_RELEASE error
If the UTS_RELEASE value cannot be determined by the configure script,
chances are the package that provides this value is not installed on the
user's system. On a RHEL6 based system, the kernel-devel package
provides this value. Thus, a hint was added to the error message to
point the user at the kernel-devel package.