Add '--with-config' configure option
Added the '--with-config' configure option to allow the user to build
srpm packages on systems that would fail a full kernel configure step.
The option takes a parameter of either 'srpm' or 'kernel'.

As a side effect, the USE_QUILT AM_CONDITIONAL had to be moved into the
LDISKFS_AC_QUILT macro to ensure the '--disable-quilt' works correctly.