Move *.m4 and *.am files into 'config' directory
The files from the 'm4' and 'am' directories were all moved into a new
'config' directory. This better aligns with the structure found in the
ZFS on Linux project, and keeps these files consolidated in a single
directory. The Makefile.am, autogen.sh, and configure.ac files were all
updated to reflect this change.