Add the WALK_SPACE_HAS_DATA_SEM macro on RHEL6
In order for Lustre to correctly build against this package, it needs
some insight into whether the ext4_ext_walk_space function which ldiskfs
was built against takes an i_data_sem internally. Since this is true on
RHEL6 based systems, a patch was added to the 2.6-rhel6 series to
introduce the macro which Lustre is expecting.