Merge LDISKFS_AC_QUILT into LDISKFS_AC_PATH_PROGS
The USE_QUILT AM_CONDITIONAL was not properly getting set by default if
a proper path to the quilt program could not be found. Thus, the
LDISKFS_AC_QUILT macro internals had to be merged into the
LDISKFS_AC_PATH_PROGS macro. This allows the AM_CONDITIONAL to be set
according the $QUILT and $enable_quilt variables.